Learned about website scraping. This udpates the code so that I don't accidently scarpe something i'm not suppsoed to.
I assume a one off accidental scrape is allowed and they won't care.

It seems to for data farmers with copyrighted material.

I shouldn't run into that issue anyways anymore

All in all this was a good mini demo on web scraping.

I'm sure I can use it for a lot of other things such as monitoring a blog forum for spam/malicious content